added gawk datamash unzip unrar
If you are like many computer users, you would frequently like to make changes in various text files wherever certain patterns appear, or extract data from parts of certain lines while discarding the rest. To write a program to do this in a language such as C or Pascal is a time-consuming inconvenience that may take many lines of code. The job is easy with awk, especially the GNU implementation: gawk.
https://www.gnu.org/software/gawk/

GNU datamash is a command-line program which performs basic numeric, textual and statistical operations on input textual data files.

https://www.gnu.org/software/datamash/